+++ src/main/java/http/request/QueryString.java
@@ -1,4 +1,4 @@
-package http;
+package http.request;
 
 import java.util.HashMap;
 import java.util.Map;
