+++ src/main/java/http/response/HttpResponse200.java
@@ -1,8 +1,6 @@
 package http.response;
 
-import http.HttpResponseCode;
-
-public class HttpResponse200 extends HttpResponse {
+public class HttpResponse200 extends HttpResponse_ {
 
     public HttpResponse200(String version) {
         super(HttpResponseCode.OK, version);
