+++ src/main/java/http/response/HttpResponse302.java
@@ -1,8 +1,6 @@
 package http.response;
 
-import http.HttpResponseCode;
-
-public class HttpResponse302 extends HttpResponse {
+public class HttpResponse302 extends HttpResponse_ {
 
     public HttpResponse302(String path, String version) {
         super(HttpResponseCode.FOUND, version);
