+++ src/test/java/http/ProtocolTest.java
@@ -1,5 +1,6 @@
 package http;
 
+import http.request.Protocol;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
